% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiCor_Peaks.R
\name{Licor_Peak}
\alias{Licor_Peak}
\title{A Li-Cor peak finding function}
\usage{
Licor_Peak(cut.off = 3)
}
\arguments{
\item{cut.off}{Cut off value used to define when the peak should start and end. Defaults to 3.}
}
\description{
This function provides user's of a Li-Cor Analyzer a simple way to extract and compile peak results such as peak CO2 emissions data.
It will loop through a folder containing the .txt files that the Li-Cor Analyzer outputs, and compiles the results into a single R dataframe.
the output of the this dataframe will give the user:
the filename that the sample belongs too,
the sample name and replicate number,
the sum of the peak range,
and the maximum peak value in each range.
}
\examples{
setwd(path.package("LiCor.Peaks"))
output <- Licor_Peak()
}
\keyword{CO2,}
\keyword{Li-Cor}
\keyword{LiCor,}
\keyword{Peak,}
