% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiCor_Package.R
\name{Licor_Peaks}
\alias{Licor_Peaks}
\title{A Li-Cor peak finding function}
\usage{
Licor_Peaks(cut.off = 2, standard.sum = F)
}
\arguments{
\item{cut.off}{Cut off value used to define when the peak should start and end. Defaults to 2.}

\item{standard.sum}{Specifies whether or not the user wants a summary of the standard statistics in the output}
}
\description{
This function provides user's of a Li-Cor Analyzer a simple way to extract and compile peak results such as peak CO2 emissions data.
It will loop through a folder containing the .txt files that the Li-Cor Analyzer outputs, and compiles the results into a single R dataframe.
the output of the this dataframe will give the user:
the filename that the sample belongs too,
the sample name and replicate number,
the sum of the peak range,
and the maximum peak value in each range.
}
\examples{
setwd(path.package("licor.peaks"))
output <- Licor_Peaks()
}
\keyword{CO2,}
\keyword{Li-Cor}
\keyword{LiCor,}
\keyword{Peak,}
