% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiCor_Package.R
\name{Plot.Licor}
\alias{Plot.Licor}
\title{A Plotting function used with output from Li-Cor Peak finding function}
\usage{
Plot.Licor(data, file = NULL, sample = NULL, std.curve = F)
}
\arguments{
\item{data}{output from the LiCor_Peaks() function}

\item{file}{the name of the file the user wishes to look at}

\item{sample}{the specific sample the user wishes to look at}

\item{curve}{a logical argument that states whether the user wishes to look at the standard curve or not}
}
\description{
This function provides a plotting framework for exploring the data output from the LiCor_Peaks() function
}
\examples{
setwd(path.package("licor.peaks"))
output <- Licor_Peaks()
Plot.Licor(output, curve = T)
Plot.Licor(output, file = "vn_clear_071621.txt", curve = T)
Plot.Licor(output, file = "vn_clear_07292021.txt", curve = T)
Plot.Licor(output, file = "vn_veg_07292021.txt", curve = T)
Plot.Licor(output, file = "vn_darkveg_071621.txt")
Plot.Licor(output, sample = "NN_DARKVEG")
Plot.Licor(output, file = "vn_darkveg_071621.txt", sample = "NC_DARKVEG")
}
\keyword{CO2,}
\keyword{Li-Cor}
\keyword{LiCor,}
\keyword{Peak,}
